+++ presto-main/src/main/java/com/facebook/presto/operator/aggregation/TypedSet.java
@@ -90,7 +90,7 @@ public class TypedSet
             containsNullElement = true;
         }
         else {
-            int hashPosition = getHashPositionOfElement(block, position);
+            long hashPosition = getHashPositionOfElement(block, position);
             if (blockPositionByHash.get(hashPosition) == EMPTY_SLOT) {
                 addNewElement(hashPosition, block, position);
             }
@@ -110,9 +110,9 @@ public class TypedSet
     /**
      * Get slot position of element at {@code position} of {@code block}
      */
-    private int getHashPositionOfElement(Block block, int position)
+    private long getHashPositionOfElement(Block block, int position)
     {
-        int hashPosition = getMaskedHash(hashPosition(elementType, block, position));
+        long hashPosition = getMaskedHash(hashPosition(elementType, block, position));
         while (true) {
             int blockPosition = blockPositionByHash.get(hashPosition);
             // Doesn't have this element
@@ -128,7 +128,7 @@ public class TypedSet
         }
     }
 
-    private void addNewElement(int hashPosition, Block block, int position)
+    private void addNewElement(long hashPosition, Block block, int position)
     {
         elementType.appendTo(block, position, elementBlock);
         if (elementBlock.getSizeInBytes() > FOUR_MEGABYTES) {
@@ -171,7 +171,7 @@ public class TypedSet
         return maxFill;
     }
 
-    private int getMaskedHash(int rawHash)
+    private long getMaskedHash(long rawHash)
     {
         return rawHash & hashMask;
     }
