+++ presto-spi/src/main/java/com/facebook/presto/spi/type/VarcharType.java
@@ -89,7 +89,7 @@ public final class VarcharType
     }
 
     @Override
-    public int hash(Block block, int position)
+    public long hash(Block block, int position)
     {
         return block.hash(position, 0, block.getLength(position));
     }
